chore: misc improvements
