+++ .env.example
@@ -28,6 +28,7 @@
 # REDIS_PASSWORD=cow_password
 
 # Database
+# DATABASE_HOST=localhost
 # DATABASE_USERNAME=bff-db-user
 # DATABASE_PASSWORD=bff-db-password
 # DATABASE_NAME=bff-db
