+++ libs/shared/src/types.ts
@@ -1,3 +1,5 @@
+import pino from 'pino';
+
 // TODO: Import from SDK
 export enum SupportedChainId {
   MAINNET = 1,
@@ -6,3 +8,5 @@ export enum SupportedChainId {
   ARBITRUM_ONE = 42161,
   SEPOLIA = 11155111,
 }
+
+export type Logger = pino.Logger;
